requirements.txt
boto3>=1.4.7
dj-database-url>=0.4.2
dj-email-url>=0.0.10
#django-danceschool>=0.9.0
-e hg+https://danceschoolrouser:Q3$)cYG=W=FXwV.B@novokhatko.pp.ua/scm/hg/django-danceschool#egg=django_danceschool
#-e file://data/web/src/django-danceschool#egg=django_danceschool
django-redis>=4.10.0
django-storages>=1.6.5
gunicorn>=19.6.0
paypalrestsdk>=1.12.0
peewee>=2.10.1
psycopg2-binary>=2.8.6
redis>=2.10.6
squareconnect>=2.20180712.3
stripe>=1.62.0
whitenoise>=3.2
